requirements.txt
streamlit
pandas
plotly
numpy
fpdf2>=2.8.5